requirements.txt
streamlit
requests
python-dotenv
boto3